requirements.txt
pandas
scikit-allel
unipressed
mygene
pyfaidx